api/requirements.txt
# インストールパッケージ
Django==3.2.4

djangorestframework==3.12.4

mysqlclient==2.0.3

openpyxl==3.0.9

PyJWT==1.7.1

django-cors-headers==3.13.0